refactor(api): move course index to src
