+++ api/lib/routes.js
@@ -14,7 +14,6 @@ import * as certificationPointOfContacts from './application/certification-point
 import * as certifications from './application/certifications/index.js';
 import * as complementaryCertificationCourseResults from './application/complementary-certification-course-results/index.js';
 import * as countries from './application/countries/index.js';
-import * as courses from './application/courses/index.js';
 import * as frameworks from './application/frameworks/index.js';
 import * as healthcheck from './application/healthcheck/index.js';
 import * as lcms from './application/lcms/index.js';
@@ -49,7 +48,6 @@ const routes = [
   certifications,
   complementaryCertificationCourseResults,
   countries,
-  courses,
   featureToggles,
   healthcheck,
   lcms,
