+++ api/src/shared/routes.js
@@ -1,6 +1,7 @@
 import * as assessmentsRoutes from './application/assessments/index.js';
 import * as challengesRoutes from './application/challenges/index.js';
+import * as courses from './application/courses/index.js';
 
-const sharedRoutes = [assessmentsRoutes, challengesRoutes];
+const sharedRoutes = [assessmentsRoutes, challengesRoutes, courses];
 
 export { sharedRoutes };
